editoast: refactor persist_batch through a custom derive
