+++ editoast/editoast_derive/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "editoast_derive"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+syn = "1.0"
+quote = "1.0"
+proc-macro2 = "1.0"
+darling = "0.14.2"
+
+[lib]
+proc-macro=true
