+++ editoast/src/schema/buffer_stop.rs
@@ -4,16 +4,15 @@ use super::OSRDIdentified;
 
 use super::OSRDTyped;
 use super::ObjectType;
-use crate::api_error::ApiError;
 use crate::infra_cache::{Cache, ObjectCache};
 use derivative::Derivative;
 use diesel::sql_types::{Double, Text};
-use diesel::ExpressionMethods;
-use diesel::{PgConnection, RunQueryDsl};
+use editoast_derive::Model;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq)]
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Model)]
 #[serde(deny_unknown_fields)]
+#[model(table = "crate::tables::osrd_infra_bufferstopmodel")]
 #[derivative(Default)]
 pub struct BufferStop {
     #[derivative(Default(value = r#"generate_id("buffer_stop")"#))]
@@ -22,35 +21,6 @@ pub struct BufferStop {
     pub track: String,
     pub position: f64,
     pub applicable_directions: ApplicableDirections,
-}
-
-impl BufferStop {
-    pub fn persist_batch(
-        values: &[Self],
-        infrastructure_id: i32,
-        conn: &PgConnection,
-    ) -> Result<(), Box<dyn ApiError>> {
-        use crate::tables::osrd_infra_bufferstopmodel::dsl::*;
-
-        let datas = values
-            .iter()
-            .map(|value| {
-                (
-                    obj_id.eq(value.get_id().clone()),
-                    data.eq(serde_json::to_value(value).unwrap()),
-                    infra_id.eq(infrastructure_id),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        for data_chunk in datas.chunks(65534) {
-            diesel::insert_into(osrd_infra_bufferstopmodel)
-                .values(data_chunk)
-                .execute(conn)?;
-        }
-
-        Ok(())
-    }
 }
 
 impl OSRDTyped for BufferStop {
