+++ editoast/src/schema/catenary.rs
@@ -1,54 +1,25 @@
 use super::generate_id;
 use super::ApplicableDirectionsTrackRange;
 use super::OSRDIdentified;
+use editoast_derive::Model;
 
 use super::OSRDTyped;
 use super::ObjectType;
-use crate::api_error::ApiError;
-use crate::diesel::ExpressionMethods;
-use crate::diesel::RunQueryDsl;
+
 use crate::infra_cache::Cache;
 use crate::infra_cache::ObjectCache;
 use derivative::Derivative;
-use diesel::PgConnection;
+
 use serde::{Deserialize, Serialize};
-#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq)]
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Model)]
 #[serde(deny_unknown_fields)]
+#[model(table = "crate::tables::osrd_infra_catenarymodel")]
 #[derivative(Default)]
 pub struct Catenary {
     #[derivative(Default(value = r#"generate_id("catenary")"#))]
     pub id: String,
     pub voltage: f64,
     pub track_ranges: Vec<ApplicableDirectionsTrackRange>,
-}
-
-impl Catenary {
-    pub fn persist_batch(
-        values: &[Self],
-        infrastructure_id: i32,
-        conn: &PgConnection,
-    ) -> Result<(), Box<dyn ApiError>> {
-        use crate::tables::osrd_infra_catenarymodel::dsl::*;
-
-        let datas = values
-            .iter()
-            .map(|value| {
-                (
-                    obj_id.eq(value.get_id().clone()),
-                    data.eq(serde_json::to_value(value).unwrap()),
-                    infra_id.eq(infrastructure_id),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        for data_chunk in datas.chunks(65534) {
-            diesel::insert_into(osrd_infra_catenarymodel)
-                .values(data_chunk)
-                .execute(conn)?;
-        }
-
-        Ok(())
-    }
 }
 
 impl OSRDTyped for Catenary {
