+++ editoast/src/schema/operational_point.rs
@@ -3,17 +3,17 @@ use super::OSRDIdentified;
 
 use super::OSRDTyped;
 use super::ObjectType;
-use crate::api_error::ApiError;
-use crate::diesel::ExpressionMethods;
-use crate::diesel::RunQueryDsl;
+
 use crate::infra_cache::Cache;
 use crate::infra_cache::ObjectCache;
 use derivative::Derivative;
-use diesel::PgConnection;
+
+use editoast_derive::Model;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq)]
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Model)]
 #[serde(deny_unknown_fields)]
+#[model(table = "crate::tables::osrd_infra_operationalpointmodel")]
 #[derivative(Default)]
 pub struct OperationalPoint {
     #[derivative(Default(value = r#"generate_id("operational_point")"#))]
@@ -21,34 +21,6 @@ pub struct OperationalPoint {
     pub parts: Vec<OperationalPointPart>,
     #[serde(default)]
     pub extensions: OperationalPointExtensions,
-}
-
-impl OperationalPoint {
-    pub fn persist_batch(
-        values: &[Self],
-        infrastructure_id: i32,
-        conn: &PgConnection,
-    ) -> Result<(), Box<dyn ApiError>> {
-        use crate::tables::osrd_infra_operationalpointmodel::dsl::*;
-        let datas = values
-            .iter()
-            .map(|value| {
-                (
-                    obj_id.eq(value.get_id().clone()),
-                    data.eq(serde_json::to_value(value).unwrap()),
-                    infra_id.eq(infrastructure_id),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        for data_chunk in datas.chunks(65534) {
-            diesel::insert_into(osrd_infra_operationalpointmodel)
-                .values(data_chunk)
-                .execute(conn)?;
-        }
-
-        Ok(())
-    }
 }
 
 #[derive(Debug, Default, Clone, Deserialize, Serialize, PartialEq, Eq)]
