+++ editoast/src/schema/route.rs
@@ -5,18 +5,18 @@ use super::OSRDIdentified;
 use super::OSRDTyped;
 use super::ObjectType;
 use super::Waypoint;
-use crate::api_error::ApiError;
-use crate::diesel::ExpressionMethods;
-use crate::diesel::RunQueryDsl;
+
 use crate::infra_cache::Cache;
 use crate::infra_cache::ObjectCache;
-use crate::tables::osrd_infra_routemodel::dsl::*;
+
 use derivative::Derivative;
-use diesel::PgConnection;
+
+use editoast_derive::Model;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq)]
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Model)]
 #[serde(deny_unknown_fields)]
+#[model(table = "crate::tables::osrd_infra_routemodel")]
 #[derivative(Default)]
 pub struct Route {
     #[derivative(Default(value = r#"generate_id("route")"#))]
@@ -25,33 +25,6 @@ pub struct Route {
     pub exit_point: Waypoint,
     pub release_detectors: Vec<String>,
     pub path: Vec<DirectionalTrackRange>,
-}
-
-impl Route {
-    pub fn persist_batch(
-        values: &[Self],
-        infrastructure_id: i32,
-        conn: &PgConnection,
-    ) -> Result<(), Box<dyn ApiError>> {
-        let datas = values
-            .iter()
-            .map(|value| {
-                (
-                    obj_id.eq(value.get_id().clone()),
-                    data.eq(serde_json::to_value(value).unwrap()),
-                    infra_id.eq(infrastructure_id),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        for data_chunk in datas.chunks(65534) {
-            diesel::insert_into(osrd_infra_routemodel)
-                .values(data_chunk)
-                .execute(conn)?;
-        }
-
-        Ok(())
-    }
 }
 
 impl OSRDTyped for Route {
