+++ editoast/src/schema/track_section.rs
@@ -5,19 +5,19 @@ use super::OSRDIdentified;
 use super::OSRDTyped;
 use super::ObjectType;
 use super::TrackEndpoint;
-use crate::api_error::ApiError;
+
 use crate::chartos::BoundingBox;
-use crate::diesel::ExpressionMethods;
-use crate::diesel::RunQueryDsl;
 use crate::infra_cache::Cache;
 use crate::infra_cache::ObjectCache;
-use crate::tables::osrd_infra_tracksectionmodel::dsl::*;
+
 use derivative::Derivative;
-use diesel::PgConnection;
+
+use editoast_derive::Model;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq)]
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Model)]
 #[serde(deny_unknown_fields)]
+#[model(table = "crate::tables::osrd_infra_tracksectionmodel")]
 #[derivative(Default)]
 pub struct TrackSection {
     #[derivative(Default(value = r#"generate_id("track_section")"#))]
@@ -30,33 +30,6 @@ pub struct TrackSection {
     pub sch: LineString,
     #[serde(default)]
     pub extensions: TrackSectionExtensions,
-}
-
-impl TrackSection {
-    pub fn persist_batch(
-        values: &[Self],
-        infrastructure_id: i32,
-        conn: &PgConnection,
-    ) -> Result<(), Box<dyn ApiError>> {
-        let datas = values
-            .iter()
-            .map(|value| {
-                (
-                    obj_id.eq(value.get_id().clone()),
-                    data.eq(serde_json::to_value(value).unwrap()),
-                    infra_id.eq(infrastructure_id),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        for data_chunk in datas.chunks(65534) {
-            diesel::insert_into(osrd_infra_tracksectionmodel)
-                .values(data_chunk)
-                .execute(conn)?;
-        }
-
-        Ok(())
-    }
 }
 
 #[derive(Debug, Clone, Default, Deserialize, Serialize, PartialEq, Eq)]
