+++ editoast/src/schema/track_section_link.rs
@@ -4,51 +4,24 @@ use super::OSRDIdentified;
 use super::OSRDTyped;
 use super::ObjectType;
 use super::TrackEndpoint;
-use crate::api_error::ApiError;
-use crate::diesel::ExpressionMethods;
-use crate::diesel::RunQueryDsl;
+
 use crate::infra_cache::Cache;
 use crate::infra_cache::ObjectCache;
-use crate::tables::osrd_infra_tracksectionlinkmodel::dsl::*;
+
 use derivative::Derivative;
-use diesel::PgConnection;
+
+use editoast_derive::Model;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Eq)]
+#[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Eq, Model)]
 #[serde(deny_unknown_fields)]
+#[model(table = "crate::tables::osrd_infra_tracksectionlinkmodel")]
 #[derivative(Default)]
 pub struct TrackSectionLink {
     #[derivative(Default(value = r#"generate_id("track_section_link")"#))]
     pub id: String,
     pub src: TrackEndpoint,
     pub dst: TrackEndpoint,
-}
-
-impl TrackSectionLink {
-    pub fn persist_batch(
-        values: &[Self],
-        infrastructure_id: i32,
-        conn: &PgConnection,
-    ) -> Result<(), Box<dyn ApiError>> {
-        let datas = values
-            .iter()
-            .map(|value| {
-                (
-                    obj_id.eq(value.get_id().clone()),
-                    data.eq(serde_json::to_value(value).unwrap()),
-                    infra_id.eq(infrastructure_id),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        for data_chunk in datas.chunks(65534) {
-            diesel::insert_into(osrd_infra_tracksectionlinkmodel)
-                .values(data_chunk)
-                .execute(conn)?;
-        }
-
-        Ok(())
-    }
 }
 
 impl OSRDTyped for TrackSectionLink {
